Update Submit Paper & My Submissions connection

+++ backend/src/main/java/com/webcrafters/confease_backend/model/Keyword.java
@@ -8,12 +8,12 @@ public class Keyword {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer keyword_id;
+    private Long keyword_id;
     private String keyword;
 
     // Getters and Setters
-    public Integer getKeyword_id() { return keyword_id; }
-    public void setKeyword_id(Integer keyword_id) { this.keyword_id = keyword_id; }
+    public Long getKeyword_id() { return keyword_id; }
+    public void setKeyword_id(Long keyword_id) { this.keyword_id = keyword_id; }
 
     public String getKeyword() { return keyword; }
     public void setKeyword(String keyword) { this.keyword = keyword; }
