+++ backend/src/main/java/com/webcrafters/confease_backend/service/KeywordService.java
@@ -2,5 +2,5 @@ package com.webcrafters.confease_backend.service;
 
 import com.webcrafters.confease_backend.model.Keyword;
 
-public interface KeywordService extends CrudService<Keyword, Integer> {
+public interface KeywordService extends CrudService<Keyword, Long> {
 }
